Refactored devices queries
