+++ libs/devices/common/feature/src/lib/devices-initializer/devices-initializer.tsx
@@ -4,7 +4,7 @@
  */
 
 import { FunctionComponent, memo } from "react"
-import { useDevicesQuery } from "../queries"
+import { useDevicesQuery } from "../hooks"
 import { useApiDeviceInitializer } from "./use-api-device-initializer"
 import { useDevicesListener } from "./use-devices-listener"
 import { useHarmonyInitializer } from "./use-harmony-initializer"
