+++ libs/devices/common/feature/src/lib/hooks/devices-query-keys.ts
