+++ libs/devices/common/feature/src/lib/hooks/index.ts
