+++ libs/devices/common/feature/src/lib/hooks/use-active-device.query.ts
