+++ libs/devices/common/feature/src/lib/hooks/use-device-config.query.ts
