+++ libs/devices/common/feature/src/lib/hooks/use-device-menu.query.ts
