+++ libs/devices/common/feature/src/lib/hooks/use-device-metadata.query.ts
