+++ libs/devices/common/feature/src/lib/hooks/use-device-status.query.ts
