+++ libs/devices/common/feature/src/lib/hooks/use-devices.query.ts
